Moved the generic axios API setup to api.js and refactored the file and folder it was in to be pollutantData.js and utils respectively.
