+++ dashboard/src/components/Pollutants/PollutantList.js
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from "react";
 
 import "./PollutantList.css";
-import { getAllPollutants } from "../../Services/APIService";
+import { getAllPollutants } from "../../api.js";
 
 const PollutantList = ({ onPollutantSelect }) => {
   const [selectedPollutant, setSelectedPollutant] = useState("");
@@ -36,20 +36,7 @@ const PollutantList = ({ onPollutantSelect }) => {
               :
               pollutant.name.replace("_", " ")}({pollutant.preferredUnit.replace("_", " ")})
           </option>
-        ))
-        }
-        {/* <option value="humidity">Humidity</option>
-        <option value="pm1">PM1</option>
-        <option value="pm10">PM10</option>
-        <option value="pm25">PM25</option>
-        <option value="pressure">Pressure</option>
-        <option value="temperature">Temperature</option>
-        <option value="um003">UM003</option>
-        <option value="um005">UM005</option>
-        <option value="um010">UM010</option>
-        <option value="um025">UM025</option>
-        <option value="um050">UM050</option>
-        <option value="um100">UM100</option> */}
+        ))}
       </select>
     </div>
   );
